Add a _redirects file
This file allows netlify to rewrite 404 requests to always serve index.html. Doing
so allows the single page app to work. Requests to a path like /add work if the
user clicks to them. But if the browser loads that path, netlify doesn't find a file
named add.html and returns a 404.

The redirects file tells the file server to always
return index.html for any request—which is what we want since everything is client
driven.